requirements.txt
aiogram>=3.3.0,<4.0.0
APScheduler>=3.10.4
python-dotenv>=1.0.1

